feat: validation validation marque est modele obligatoir depuis la liste

+++ Back/routes/vehicules.js
@@ -7,7 +7,7 @@ const { vehiculeSchema } = require('../utils/validators');
 const { asyncHandler } = require('../utils/handler');
 
 router.post('/vehicules', verifierToken, asyncHandler(async(req, res) => {
-    
+
     const { marque, modele, annee, kilometrage, carburant, region, entretien, ct } = req.body;
     // Validation z pour le formulaire
     const result = vehiculeSchema.safeParse({
@@ -23,7 +23,7 @@ router.post('/vehicules', verifierToken, asyncHandler(async(req, res) => {
     if(!result.success){
         return res.status(400).json({ erreur: result.error.issues[0].message });
     }
-    // ke middleware verifiertoken a deja decodé le token
+    // le middleware verifiertoken a deja decodé le token
     const user_id = req.user.id;
     //  On Verifie le nombre de vehicule avant d'ajouter
     const count = await sql`SELECT COUNT(*) FROM vehicules WHERE user_id = ${user_id}`;
@@ -33,7 +33,7 @@ router.post('/vehicules', verifierToken, asyncHandler(async(req, res) => {
         return res.status(400).json({ erreur: 'Vous avez atteint la limite de 5 véhicules !' })
     }
 
-    const ajouterVehicule = await sql`INSERT INTO vehicules ( user_id, marque, modele, annee, kilometrage, carburant, region, entretien, ct ) 
+    const ajouterVehicule = await sql`INSERT INTO vehicules ( user_id, marque, modele, annee, kilometrage, carburant, region, entretien, ct )
     VALUES (${user_id},${marque}, ${modele}, ${annee}, ${kilometrage}, ${carburant}, ${region}, ${entretien}, ${ct}) RETURNING *`;
     res.json({ message: 'Vehicule ajoute avec succes ✅', vehicule: ajouterVehicule[0]});
 }));
@@ -43,8 +43,8 @@ router.get('/vehicules', verifierToken, asyncHandler(async(req, res) => {
     const user_id = req.user.id;
 
     const vehicules = await sql`
-        SELECT vehicules.*, scores.score_global 
-        FROM vehicules 
+        SELECT vehicules.*, scores.score_global
+        FROM vehicules
         LEFT JOIN scores ON vehicules.id = scores.vehicule_id
         WHERE vehicules.user_id = ${user_id}
     `;
@@ -52,8 +52,9 @@ router.get('/vehicules', verifierToken, asyncHandler(async(req, res) => {
 }));
 
 router.delete('/vehicules/:id', verifierToken, asyncHandler(async(req, res) => {
-    const id = parseInt(req.params.id);
-    const result = await sql`DELETE FROM vehicules WHERE id = ${id}`;
+    const user_id = req.user.id
+    const id = parseInt(req.params.id)
+    const result = await sql`DELETE FROM vehicules WHERE id = ${id} AND user_id = ${user_id}`;
     res.json(result);
